add web interface
This commit adds a web interface based on rails. This is just a very
preliminary version that gets the data from a database.

There is still work to do to insert that data in the database from the
gems-status gem.

The goal of having a web interface is to be able to add some interaction,
like updating the security information.