use erb for including yaml files
This way, the configuration can be loaded in one place and that means
that we can instead load an object. That is needed in order to be
able to use gems-status as an API and store configurations in
database instead of yaml files for example, like we do in the
gems-status-web.