simplify: only one source and textview instead of html
This commit simplifies the gems-status.
From now on, only one source since that is what we are actually using.
The lockfile is only one gemfile.lock. That simplifies the software.
Output in text format instead of html. That simplifies the view a lot,
and since there is the gems-status-web project, we don't need the html.